requirements.txt
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
blessed==1.20.0
boto3==1.28.57
botocore==1.31.57
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==41.0.2
curtsies==0.4.1
cwcwidth==0.1.8
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2
django-crispy-forms==2.0
django-extensions==3.2.1
django-forms-dynamic==1.0.0
django-money==3.1.0
django-smart-selects==1.6.0
django-storages==1.14.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
docopt==0.6.2
exceptiongroup==1.1.3
gunicorn==21.2.0
h11==0.14.0
httpie==3.2.2
httptools==0.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
pipdeptree==2.11.0
pipreqs==0.4.13
psycopg==3.1.12
py-moneyed==2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.30.0
requests-toolbelt==1.0.0
rich==13.5.3
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sorl-thumbnail==12.9.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
wcwidth==0.2.6
websocket-client==1.5.2
websockets==11.0.3
whitenoise==6.5.0
yarg==0.1.9
zope.interface==6.0
